+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Added a new gamerule `iguanatweaksreborn:deathLoseItems`
+  * Defaults to false, if true players' items will disappear on death
 * Sheep now have higher chance to regrow wool in cold seasons instead of hot seasons
 * Reduced red color of the red outline on blocks
 * Livestock data's `breeding_fail_chance_modifiers` are now added to the base chance instead of multiplying it